Simplify threading model
Since both threads lock the repository before doing any operations on
it, there is no need to synchronize the repository methods. The
logging functions are threadsafe, so log errors directly in thread
instead of storing in errors[] list and handle later. All this also
means that the repository lock should not be reentrant, there is
just an overall, sort of global lock only used once in each thread.