Retire long-running fetch thread, use schedule instead.
There are two problems with current thread design: the fast-polling
fetcher thread which always look for if it's shut down, and that the
poll call which only is meaningful after a fetch is done at a random
times.

New design uses schedule to run fetch regularly, and fetch schedules
poll to run directly after fetch completion. The fetcher thread is
still asynchronous, but just a one-shot one.